fix(admin): map missing progress data and enforce 100% completion rule in UI-5 Level 2

+++ lms-platform/enrollment-service/src/main/java/com/lms/enrollmentservice/adapter/in/rest/dto/AdminStudentEnrollmentResponse.java
@@ -2,6 +2,8 @@ package com.lms.enrollmentservice.adapter.in.rest.dto;
 
 import lombok.Builder;
 import lombok.Data;
+
+import java.math.BigDecimal;
 import java.time.ZonedDateTime;
 
 @Data
@@ -12,7 +14,7 @@ public class AdminStudentEnrollmentResponse {
     private String studentName;
     private String studentEmail;
     private String status;
-    private Double progressPercentage;
+    private BigDecimal progressPercentage;
     private Long lastAccessedLessonId;
     private ZonedDateTime enrolledAt;
     private ZonedDateTime completedAt;
